begin separating compositor-proxy as CLI and as lib
