+++ README.md
@@ -100,7 +100,7 @@ This creates a set of files in the `package` directory. The `run.sh` script acce
 
 	Options
 	  --help, Print this help text.
-	  --static-session-id=..., Use and accept this and only this session id when communicating.
+      --static-session-id=...,  Mandatory. Only use and accept this session id when communicating.
 	  --config-path=...,  Use a configuration file located at this file path.
 
 	Examples
